Updated science club invoicing to typescript
Consumed the shared types in both functions and react.
Updated confirmation dialog to use generics to better handle merging of props.
Moved InvoiceStatus into its own file and renamed as InvoiceStatusWithAction so as not to conflict with the InvoiceStatus interface defined in fizz-kidz.
