+++ firebase/functions/src/acuity/shared/index.ts
@@ -1,4 +1,4 @@
-import * as Acuity from '../../../types/acuity'
+import { Acuity } from 'fizz-kidz'
 
 export function isAcuityError(object: any | Acuity.Error): object is Acuity.Error {
     return (object as Acuity.Error).error !== undefined
