+++ firebase/src/utilities/firebase/functions/index.ts
@@ -1,0 +1,2 @@
+// https://stackoverflow.com/a/51507473
+export type FunctionsResult<T> = Omit<firebase.functions.HttpsCallableResult, 'data'> & { data: T }
